requirements.txt
PyGithub==1.44.1
PyQt5==5.8.2
PyQt5-sip==4.19.14
